ZippedDataをUtility化。不要なメソッドを削除
